Atualização de métodos

+++ src/main/java/io/github/vrperdomo/libaryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.github.vrperdomo.libaryapi.service;
 
+import io.github.vrperdomo.libaryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.vrperdomo.libaryapi.model.Autor;
 import io.github.vrperdomo.libaryapi.repository.AutorRepository;
+import io.github.vrperdomo.libaryapi.repository.LivroRepository;
+import io.github.vrperdomo.libaryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,12 +15,17 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository autorRepository;
+    private final AutorValidator autorValidator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository autorRepository) {
+    public AutorService(AutorRepository autorRepository, AutorValidator autorValidator, LivroRepository livroRepository) {
         this.autorRepository = autorRepository;
+        this.autorValidator  = autorValidator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
+        autorValidator.validarAutor(autor);
         return autorRepository.save(autor);
     }
 
@@ -26,6 +34,10 @@ public class AutorService {
     }
 
     public void deletarAutor(Autor autor) {
+
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não é perdimitido excluir um autor que possui livros cadastrados!");
+        }
         autorRepository.delete(autor);
     }
 
@@ -49,9 +61,15 @@ public class AutorService {
 
     public void atualizar(Autor autor) {
 
-        if(autor.getId() == null) {
+        if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar, é necessário que o autor já esteja salvo na base");
         }
-         autorRepository.save(autor);
+
+        autorValidator.validarAutor(autor);
+        autorRepository.save(autor);
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
